requirements.txt
# ------------------ GraphQL and API ------------------
ariadne==0.23.0           # Schema-first GraphQL framework
fastapi==0.115.0          # High-performance ASGI web framework
uvicorn==0.30.0           # ASGI server to run FastAPI app

# ------------------ Geometry & BIM ------------------
ifcopenshell==0.8.0       # IFC parsing and geometry access (OpenCascade-based)
pythonocc-core==7.7.2     # OpenCascade bindings for 3D geometry operations
shapely==2.0.6            # WKT & 2D geometry library (area, perimeter, intersection)

# ------------------ Semantic/RDF (GeoSPARQL) ------------------
rdflib==7.0.0             # RDF triple management and SPARQL querying

# ------------------ Lifecycle & Scientific ------------------
numpy==2.1.1              # Numerical support (can assist volume, matrix ops)

# ------------------ Security & Authentication ------------------
PyJWT==2.9.0              # JWT token generation for user authentication

# ------------------ Testing ------------------
pytest==8.2.2             # Test framework

# ------------------ Optional Visualization ------------------
matplotlib==3.9.2         # Visualize geometry (if needed)
contourpy==1.3.0          # Dependency of matplotlib
cycler==0.12.1            # Style cycling for plots
fonttools==4.54.1         # Font support for matplotlib
